Fix layout when scrollbars on.
When scrollbars are turned on in system settings, and the page needed
vertical scrolling, the vertical scroll bar was not shrinking the
content, which meant that the content had a horizontal scroll bar as
well, and the vertical scroll bar clipped the right edge of the content.
This is now fixed.